shared/linux: remove padding from union bpf_attr
Clang does not initialize the padding of unions. Linux expects memory
after the last element of the union and up until sizeof(bpf_attr) to be
zero. This is not always the case, when the padding of the union is not
explicitly initialized to zero. Add a dummy variable to the union
bpf_attr to remove padding from it, thereby avoiding this issue.

See also https://reviews.llvm.org/D68115